Refactor: ShopSearchコンポーネントのショップ情報を表示するコードをShopCardコンポーネントとして切り分け
